+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Unreleased
-    - Gaussian MLE now has automatic and adaptive threshold
     - Added early stopping window to AdaBoost
+    - Gaussian MLE now has automatic and adaptive threshold
+    - Loda now has automatic and adaptive threshold
     - Added params method to Estimator and Embedder interface
     - t-SNE now compatible with categorical distance kernels
     - Grid Search implements the Wrapper interface